02_finetuning/pyproject.toml
[project]
name = "mistral-dbt-finetuning"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["openai>=1.0.0", "python-dotenv>=1.0.0"]

[tool.uv]

[dependency-groups]
dev = []
